+++ src/tu/one_sum.h
@@ -14,14 +14,17 @@ extern "C" {
 
 typedef struct
 {
-  TU_MATRIX* matrix;      /**< \brief Sparse matrix. */
-  TU_MATRIX* transpose;   /**< \brief Sparse transposed matrix. */
-  int* rowsToOriginal;    /**< \brief Maps component rows to original matrix rows. */
-  int* columnsToOriginal; /**< \brief Maps component columns to original matrix columns. */
+  TU_MATRIX* matrix;          /**< \brief Sparse matrix. */
+  TU_MATRIX* transpose;       /**< \brief Sparse transposed matrix. */
+  size_t* rowsToOriginal;     /**< \brief Maps component rows to original matrix rows. */
+  size_t* columnsToOriginal;  /**< \brief Maps component columns to original matrix columns. */
 } TU_ONESUM_COMPONENT;
 
 /**
  * \brief Decomposes int matrix into 1-connected submatrices.
+ *
+ * Allocates an array \p components with an entry per 1-connected submatrix. The caller has to free this array and
+ * its members.
  */
 
 TU_ERROR decomposeOneSum(
@@ -29,12 +32,12 @@ TU_ERROR decomposeOneSum(
   TU_MATRIX* matrix,                /**< Matrix */
   size_t matrixType,                /**< Size of base type of matrix. */
   size_t targetType,                /**< Size of base type of component matrices. */
-  int* pnumComponents,               /**< Number of components */
-  TU_ONESUM_COMPONENT** components, /**< Component information */
-  int* rowsToComponents,            /**< Mapping of rows of \p matrix to components. Can be \c NULL. */
-  int* columnsToComponents,         /**< Mapping of columns of \p matrix to components. Can be \c NULL. */
-  int* rowsToComponentRows,         /**< Mapping of rows to rows of the component. Can be \c NULL. */
-  int* columnsToComponentColumns    /**< Mapping of columns to columns of the component. Can be \c NULL. */
+  size_t* pnumComponents,           /**< Pointer for storing the number of components. */
+  TU_ONESUM_COMPONENT** components, /**< Pointer for storing the array with component information. */
+  size_t* rowsToComponents,         /**< Mapping of rows of \p matrix to components (may be \c NULL). */
+  size_t* columnsToComponents,      /**< Mapping of columns of \p matrix to components (may be \c NULL). */
+  size_t* rowsToComponentRows,      /**< Mapping of rows to rows of the component (may be \c NULL). */
+  size_t* columnsToComponentColumns /**< Mapping of columns to columns of the component (may be \c NULL). */
 );
 
 #ifdef __cplusplus
